+++ lib/src/rect_clipper.dart
@@ -1,6 +1,6 @@
 import 'dart:ui';
 import 'package:flutter/material.dart';
-import 'package:flutter_animated_button/flutter_animated_button.dart';
+import 'package:flutter_animated_button/src/transition_type.dart';
 
 class RectClipper extends CustomClipper<Path> {
   final double clipFactor;