+++ CHANGELOG.md
@@ -1,3 +1,17 @@
-## [0.0.1] - TODO: Add release date.
+## [0.2.1]   01-Nov-2020  :rocket:
 
-* TODO: Describe initial release.
+* Bug fixed
+
+## [0.2.0]   01-Nov-2020  :rocket:
+
+* Improve document quality.
+* Add ```enable``` button feature
+* Add ```onChange``` funcion which return status of button (```true``` - selected , ```false``` - Deselected)
+* Bug fixed
+
+## [0.1.0]   31-Oct-2020 :package: :rocket:
+
+* Initial Release.
+* Provide basic slide animation with customization of color, sizes and textStyle.
+* Provide AnimationController for get status of animation and handle animation 
+